+++ bin/commands/find.rb
@@ -7,7 +7,7 @@ class App
   (partial matches allowed). Add a + before a token to make it required, e.g. `na find +feature +maybe`,
   add a - or ! to ignore matches containing that token.'
   arg_name 'PATTERN'
-  command %i[find grep] do |c|
+  command %i[find grep search] do |c|
     c.example 'na find feature idea swift', desc: 'Find all actions containing feature, idea, and swift'
     c.example 'na find feature idea -swift', desc: 'Find all actions containing feature and idea but NOT swift'
     c.example 'na find -x feature idea', desc: 'Find all actions containing the exact text "feature idea"'
@@ -64,17 +64,33 @@ class App
         NA.save_search(title, "#{NA.command_line.join(' ').sub(/ --save[= ]*\S+/, '').split(' ').map { |t| %("#{t}") }.join(' ')}")
       end
 
-
       depth = if global_options[:recurse] && options[:depth].nil? && global_options[:depth] == 1
                 3
               else
                 options[:depth].nil? ? global_options[:depth].to_i : options[:depth].to_i
               end
 
+      if options[:exact] || options[:regex]
+        search = args.join(' ')
+      else
+        search = args.join(' ').gsub(/(?<=\A|[ ,])(?<req>[+\-!])?@(?<tag>[^ *=<>$\^,@(]+)(?:\((?<value>.*?)\)| *(?<op>[=<>]{1,2}|[*$\^]=) *(?<val>.*?(?=\Z|[,@])))?/) do |arg|
+          m = Regexp.last_match
+          string = if m['value']
+                     "#{m['req']}#{m['tag']}=#{m['value']}"
+                   else
+                     m[0]
+                   end
+          options[:tagged] << string.sub(/@/, '')
+          ''
+        end
+      end
+
+      search = search.gsub(/ +/, ' ').strip
+
       all_req = options[:tagged].join(' ') !~ /[+!\-]/ && !options[:or]
       tags = []
       options[:tagged].join(',').split(/ *, */).each do |arg|
-        m = arg.match(/^(?<req>[+\-!])?(?<tag>[^ =<>$\^]+?)(?:(?<op>[=<>]{1,2}|[*$\^]=)(?<val>.*?))?$/)
+        m = arg.match(/^(?<req>[+\-!])?(?<tag>[^ =<>$\^]+?) *(?:(?<op>[=<>]{1,2}|[*$\^]=) *(?<val>.*?))?$/)
 
         tags.push({
                     tag: m['tag'].wildcard_to_rx,
@@ -85,19 +101,23 @@ class App
                   })
       end
 
+      search_for_done = false
+      tags.each { |tag| search_for_done = true if tag[:tag] =~ /done/ }
+      options[:done] = true if search_for_done
+
       tokens = nil
       if options[:exact]
-        tokens = args.join(' ')
+        tokens = search
       elsif options[:regex]
-        tokens = Regexp.new(args.join(' '), Regexp::IGNORECASE)
+        tokens = Regexp.new(search, Regexp::IGNORECASE)
       else
         tokens = []
-        all_req = args.join(' ') !~ /[+!\-]/ && !options[:or]
+        all_req = search !~ /[+!\-]/ && !options[:or]
 
-        args.join(' ').split(/ /).each do |arg|
+        search.split(/ /).each do |arg|
           m = arg.match(/^(?<req>[+\-!])?(?<tok>.*?)$/)
           tokens.push({
-                        token: m['tok'],
+                        token: Regexp.escape(m['tok']),
                         required: all_req || (!m['req'].nil? && m['req'] == '+'),
                         negate: !m['req'].nil? && m['req'] =~ /[!\-]/
                       })